re-organized .gitignore, added some .gitkeep files
I like my .gitignore documented and structured in a sane way.
It can grow quite a lot and in some cases it's quite important to see
what is ignored and why. Also the order of entries matters, which is why
it shouldn't just be a bunch of lines.

Also added .gitkeep files for `var/`. Initial composer install
(post-install-cmd) failed due to the missing var directory.