I splited up the basecontroller, and took all methods out which declare a route. We need to be able to overwrite a action and change the method signature. (for example having a parameter for the grid/list actions) the current class layout hasnt allowed this.
All current controllers now extends the new CrudController - which in turn extends the BaseController.
So now you can use the basecontroller without have to worry about routes you dont need or whant to override.

If you need to extend either class - the rule is simple:
if the method defines a route it needs to go to the CrudController all other methods can be in the base controller.

This change doesnt break any backward compatibility - you just have to extend CrudController in the future.

ping @david.herrmann, @niklas.devries